backend/requirements.txt
fastapi
uvicorn[standard]
supabase
httpx
python-dotenv
pydantic
apscheduler
asyncpraw
google-genai
